requirements.txt
streamlit
pandas
ollama
folium
streamlit-folium
plotly
google-generativeai